fix(roles): user_roles column types INT (match users.id and roles.id, not BIGINT)
users.id and roles.id are both INT signed (per DESCRIBE output).
The previous BIGINT declaration triggered MySQL error 3780
'Referencing column and referenced column are incompatible'.

+++ sql/multi_role_user.sql
@@ -7,12 +7,16 @@
 --   - Existing single-role users are backfilled so user_roles always
 --     contains at least their primary role.
 --
+-- Column types match users.id and roles.id exactly so FK constraints
+-- pass. If your users.id is INT UNSIGNED, change the types accordingly
+-- (run `DESCRIBE users;` and `DESCRIBE roles;` first).
+--
 -- Idempotent. Safe to re-run.
 
 CREATE TABLE IF NOT EXISTS user_roles (
-  user_id     BIGINT NOT NULL,
-  role_id     BIGINT NOT NULL,
-  granted_by  BIGINT NULL,        -- audit: who added the grant
+  user_id     INT NOT NULL,
+  role_id     INT NOT NULL,
+  granted_by  INT NULL,           -- audit: who added the grant
   granted_at  DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (user_id, role_id),
   KEY idx_user_roles_user (user_id),
@@ -21,7 +25,6 @@ CREATE TABLE IF NOT EXISTS user_roles (
   CONSTRAINT fk_user_roles_role FOREIGN KEY (role_id) REFERENCES roles(id) ON DELETE CASCADE
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
--- Backfill from each user's current primary role. INSERT IGNORE so re-runs
--- don't duplicate rows.
+-- Backfill from each user's current primary role.
 INSERT IGNORE INTO user_roles (user_id, role_id, granted_at)
 SELECT id, role_id, NOW() FROM users WHERE role_id IS NOT NULL;
